Move getPhysicalHiveColumnHandles into HiveUtil
